[WIP] Parser strategies [Trello card](https://trello.com/c/IWr2WOQy/7-implementer-une-notion-de-strategie-de-correction-fix-github-update)
- [x] Refactored parser to allow different strategies
- [ ] Refactored fixer so that it can handle an array of comment
- [ ] Tests
