+++ src/parser/index.js
@@ -1,0 +1,33 @@
+import getGitHubEventHeader from './getGitHubEventHeader';
+import parsePullRequestReviewFactory from './parsePullRequestReview';
+import parsePullRequestReviewComment from './parsePullRequestReviewComment';
+
+export default (client, logger) => {
+    const parsers = {
+        pull_request_review_comment: parsePullRequestReviewComment,
+        pull_request_review: parsePullRequestReviewFactory(client, logger),
+    };
+
+    return request => {
+        let result;
+
+        const event = getGitHubEventHeader(request.headers);
+
+        if (!event) {
+            logger.debug('no event in github payload');
+            return [];
+        }
+
+        logger.debug('event in github payload', event);
+
+        if (event === 'ping') return [];
+
+        const parser = parsers[event];
+        if (!parser) return [];
+
+        result = parsePullRequestReviewComment(request);
+
+        logger.debug('result of github payload parsing', result);
+        return result;
+    };
+};
